+++ base/exports.jl
@@ -946,6 +946,7 @@ export
     isbits,
     isequal,
     isimmutable,
+    isimmutabletype,
     isless,
     ifelse,
     lexless,